Restructure main scraping functionality into separate submodule
